+++ src/extension/utils/createConfFile.ts
@@ -4,8 +4,8 @@
 
 import { workspace, Uri, window } from 'vscode'
 import { log, Sources } from './log'
-import { CONF_DIRECTORY_NAME, NewForm, SolidityObj } from '../types'
-import { cliVersion } from '../ScriptRunner'
+import { CONF_DIRECTORY_NAME, CvlVersion, NewForm, SolidityObj } from '../types'
+import { cvlVersion } from '../ScriptRunner'
 
 type ConfFile = {
   files?: string[]
@@ -152,7 +152,7 @@ export function newFormToConf(newForm: NewForm): string {
 
   if (newForm.specObj.specFile && newForm.solidityObj.mainContract) {
     // verify is an array in the old version, string in the new one
-    const oldCliVersion = cliVersion === 1
+    const oldCliVersion = cvlVersion === CvlVersion.cvlVersion1
 
     if (oldCliVersion) {
       config.verify = [